refactor: rename and move components for the menu bar
