+++ src/components/toolbar/toolbar.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 import InterpretationsAndDetailsToggler from './interpretations-and-details-toggler'
-import MenuBar from './menu-bar'
+import MenuBar from './menu-bar/menu-bar'
 import UpdateButton from './update-button'
 import { VisualizationTypeSelector } from './visualization-type-selector/visualization-type-selector'
 import { Toolbar as AnalyticsToolbar } from '@dhis2/analytics'
